physiological logging: Only empty indexes can be removed.

+++ core/src/main/java/com/orientechnologies/orient/core/storage/index/sbtree/multivalue/v3/OCellBTreeMultiValueV3.java
@@ -27,6 +27,7 @@ import com.orientechnologies.common.serialization.types.OByteSerializer;
 import com.orientechnologies.common.types.OModifiableLong;
 import com.orientechnologies.orient.core.config.OGlobalConfiguration;
 import com.orientechnologies.orient.core.encryption.OEncryption;
+import com.orientechnologies.orient.core.exception.ONonEmptyComponentCanNotBeRemovedException;
 import com.orientechnologies.orient.core.exception.OTooBigIndexKeyException;
 import com.orientechnologies.orient.core.id.ORID;
 import com.orientechnologies.orient.core.index.OAlwaysGreaterKey;
@@ -455,6 +456,11 @@ public final class OCellBTreeMultiValueV3<K> extends ODurableComponent implement
     try {
       acquireExclusiveLock();
       try {
+        final long size = size();
+        if (size > 0) {
+          throw new ONonEmptyComponentCanNotBeRemovedException(
+              getName() + " : Not empty index can not be deleted. Index has " + size + " records");
+        }
         deleteFile(atomicOperation, fileId);
         deleteFile(atomicOperation, nullBucketFileId);
