+++ core/src/main/java/com/orientechnologies/orient/core/storage/index/sbtree/singlevalue/v3/OCellBTreeSingleValueV3.java
@@ -27,6 +27,7 @@ import com.orientechnologies.common.serialization.types.OLongSerializer;
 import com.orientechnologies.common.serialization.types.OShortSerializer;
 import com.orientechnologies.orient.core.config.OGlobalConfiguration;
 import com.orientechnologies.orient.core.encryption.OEncryption;
+import com.orientechnologies.orient.core.exception.ONonEmptyComponentCanNotBeRemovedException;
 import com.orientechnologies.orient.core.exception.OTooBigIndexKeyException;
 import com.orientechnologies.orient.core.id.ORID;
 import com.orientechnologies.orient.core.id.ORecordId;
@@ -422,6 +423,12 @@ public final class OCellBTreeSingleValueV3<K> extends ODurableComponent implemen
     try {
       acquireExclusiveLock();
       try {
+        final long size = size();
+        if (size > 0) {
+          throw new ONonEmptyComponentCanNotBeRemovedException(
+              getName() + " : Not empty index can not be deleted. Index has " + size + " records");
+        }
+
         deleteFile(atomicOperation, fileId);
         deleteFile(atomicOperation, nullBucketFileId);
       } finally {
